backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

motor==3.6.0
pymongo==4.9.1

python-dotenv==1.0.1
pydantic[email]==2.9.2

feedparser==6.0.11
requests==2.32.3
httpx==0.27.2

beautifulsoup4==4.12.3
lxml==5.3.0

google-generativeai==0.8.3
tenacity==8.2.3

apscheduler==3.10.4

email-validator==2.2.0
python-multipart==0.0.9

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

PyJWT==2.8.0

orjson==3.10.7
pytz==2024.2

loguru==0.7.2